# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI智能助手系统",
    "summary": "AI智能助手系统API",
    "description": "\n## AI智能助手系统后端API\n\n基于FastAPI和LangChain 1.0框架构建的企业级AI智能助手系统，提供以下核心功能：\n\n### 核心功能\n\n* **智能对话** - 支持多轮对话、流式响应、对话历史管理\n* **RAG知识库** - 文档上传、向量化、智能检索问答\n* **Agent智能代理** - 工具调用、任务执行、推理过程可视化\n* **用户管理** - 注册登录、JWT认证、配额管理\n* **系统监控** - 使用统计、健康检查、Prometheus指标\n\n### 技术栈\n\n* **Web框架**: FastAPI 0.104+\n* **AI框架**: LangChain 1.0\n* **数据库**: MySQL 8.0, Redis 7.0, Chroma\n* **LLM**: 通义千问 (Tongyi Qianwen)\n\n### 认证方式\n\n大部分API端点需要JWT认证。在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 速率限制\n\n* 登录接口: 5次/分钟\n* 普通API: 100次/分钟\n* LLM调用: 20次/分钟\n\n### 错误响应格式\n\n所有错误响应遵循统一格式：\n```json\n{\n    \"error_code\": \"1001\",\n    \"message\": \"错误描述\",\n    \"request_id\": \"uuid\"\n}\n```\n    ",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "文档管理"
        ],
        "summary": "上传文档",
        "description": "上传单个文档到指定知识库。支持PDF、Word、TXT、Markdown格式，最大10MB。",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/upload-batch": {
      "post": {
        "tags": [
          "文档管理"
        ],
        "summary": "批量上传文档",
        "description": "批量上传多个文档到指定知识库。",
        "operationId": "upload_documents_batch_api_v1_documents_upload_batch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_batch_api_v1_documents_upload_batch_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "get": {
        "tags": [
          "文档管理"
        ],
        "summary": "获取文档列表",
        "description": "获取指定知识库的文档列表，支持分页。",
        "operationId": "get_documents_api_v1_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "tags": [
          "文档管理"
        ],
        "summary": "获取文档处理状态",
        "description": "获取指定文档的处理状态和进度。",
        "operationId": "get_document_status_api_v1_documents__document_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/preview": {
      "get": {
        "tags": [
          "文档管理"
        ],
        "summary": "获取文档预览",
        "description": "获取文档的前1000个字符的文本内容预览。",
        "operationId": "get_document_preview_api_v1_documents__document_id__preview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "max_chars",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000,
              "title": "Max Chars"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "delete": {
        "tags": [
          "文档管理"
        ],
        "summary": "删除文档",
        "description": "删除指定文档及其向量数据。",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchUploadResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentUploadResponse"
            },
            "type": "array",
            "title": "Documents",
            "description": "上传成功的文档列表"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors",
            "description": "上传失败的文件列表"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "BatchUploadResponse",
        "description": "批量上传响应"
      },
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "要上传的文档文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
      "Body_upload_documents_batch_api_v1_documents_upload_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "description": "要上传的文档文件列表"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_batch_api_v1_documents_upload_batch_post"
      },
...
      "DocumentListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总数"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "文档列表"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "DocumentListResponse",
        "description": "文档列表响应"
      },
...
      "DocumentUploadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "文档ID"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "文件名"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "文件大小（字节）"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "处理状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "file_size",
          "status",
          "created_at"
        ],
        "title": "DocumentUploadResponse",
        "description": "文档上传响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "消息内容"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "通用消息响应"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}